+++ app/src/features/settings/components/BillingTeam/components/BillingDetails/MyBillingTeamDetails/components/TeamPlanDetails/index.scss
