+++ app/src/features/settings/components/BillingTeam/components/BillingDetails/OtherBillingTeamDetails/components/TeamDetailsPopover/index.scss
